Update markers.js
Until may of last year, I was posting updates to my map on reddit, and so I have a lot of these locations stored (and am now able to verify them with the world download) - I'm pretty excited about having the ability to help work on a community driven version of the map. This is my first time doing anything on github so please forgive me if I've done it incorrectly, I tried to copy the same format as you've used here.  I think I've probably added 10 or 12 new locations, and corrected 3 or 4.